back/requirements.txt
fastapi
uvicorn[standard]
prisma
black
factory-boy
Faker